% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{as.data.frame.kuzu_networkx}
\alias{as.data.frame.kuzu_networkx}
\title{Convert a kuzu_networkx Object to a List of Data Frames}
\usage{
\method{as.data.frame}{kuzu_networkx}(x, ...)
}
\arguments{
\item{x}{A \code{kuzu_networkx} object from \code{as_networkx()}.}

\item{...}{Additional arguments (not used).}
}
\value{
A list containing two data frames: \code{nodes} and \code{edges}.
}
\description{
This S3 method converts the \code{reticulate}-wrapped \code{networkx} graph object
into a standard R list containing a \code{nodes} data frame and an \code{edges}
data frame.
}
\details{
This function shells out to Python via \code{reticulate} to perform the
conversion. It uses \code{pandas} and \code{networkx} to extract the node and edge
attributes into data frames that can be easily used in R. This is the
core of the conversion process.
}
