% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{as_networkx}
\alias{as_networkx}
\title{Convert a Kuzu Query Result to a NetworkX Object}
\usage{
as_networkx(query_result)
}
\arguments{
\item{query_result}{A \code{kuzu_query_result} object from \code{kuzu_execute()}.}
}
\value{
A \code{networkx.classes.graph.Graph} object (via \code{reticulate}).
}
\description{
This is a helper function that takes the result of \code{kuzu_execute()} and
calls the \code{.get_as_networkx()} method on the underlying Python object.
}
\details{
This function serves as the first step in a two-step process to convert
Kuzu graph data into R-native graph objects. It returns a \code{reticulate}
proxy object for a Python \code{networkx} graph. The class of this object is
set to allow dispatch to a corresponding \code{as.data.frame} method.
}
