% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_kuzu}
\alias{install_kuzu}
\title{Install the Kuzu Python package}
\usage{
install_kuzu()
}
\description{
This function checks if the 'kuzu' Python package is available in the current
reticulate environment. If not, it will ask for permission to install it
using \code{reticulate::py_install()}.
}
\examples{
\dontrun{
install_kuzu()
}
}
