% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuzu.R
\name{kuzu_database}
\alias{kuzu_database}
\title{Initialize a Kuzu Database}
\usage{
kuzu_database(path)
}
\arguments{
\item{path}{A string specifying the file path for the database. For an
in-memory database, use \code{":memory:"}.}
}
\value{
A Python object representing the Kuzu database instance.
}
\description{
Establishes a connection to a Kuzu database. If the database does not exist
at the specified path, it will be created.
}
\examples{
\dontrun{
# Create an in-memory database
db <- kuzu_database(":memory:")

# Create or connect to an on-disk database
db_disk <- kuzu_database("my_kuzu_db")
}
}
