% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kuzu_query}
\alias{kuzu_query}
\title{Execute a query on a Kuzu database}
\usage{
kuzu_query(connection, query)
}
\arguments{
\item{connection}{Connection object}

\item{query}{Cypher query string}
}
\value{
List with query results
}
\description{
Execute a query on a Kuzu database
}
