% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{as_networkx}
\alias{as_networkx}
\title{Convert a Kuzu Query Result to a NetworkX Object}
\usage{
as_networkx(query_result)
}
\arguments{
\item{query_result}{A \code{kuzu_query_result} object from \code{kuzu_execute()}.}
}
\value{
A \code{networkx.classes.graph.Graph} object (via \code{reticulate}).
}
\description{
This is a helper function that takes the result of \code{kuzu_execute()} and
calls the \code{.get_as_networkx()} method on the underlying Python object.
}
\details{
This function serves as the first step in a two-step process to convert
Kuzu graph data into R-native graph objects. It returns a \code{reticulate}
proxy object for a Python \code{networkx} graph. The class of this object is
set to allow dispatch to a corresponding \code{as.data.frame} method.
}
\examples{
\dontrun{
  conn <- kuzu_connection(":memory:")
  kuzu_execute(conn, "CREATE NODE TABLE Person(name STRING, age INT64, 
  PRIMARY KEY (name))")
  kuzu_execute(conn, "CREATE REL TABLE Knows(FROM Person TO Person, 
  since INT64)")
  kuzu_execute(conn, "CREATE (p:Person {name: 'Alice', age: 25})")
  kuzu_execute(conn, "CREATE (q:Person {name: 'Bob', age: 30})")
  kuzu_execute(conn, "MATCH (a:Person), (b:Person) WHERE a.name='Alice' 
  AND b.name='Bob' " \%+\%
                       "CREATE (a)-[:Knows {since: 2022}]->(b)")
  res <- kuzu_execute(conn, "MATCH (p:Person)-[k:Knows]->(q:Person) 
  RETURN p, k, q")

  # Convert to a networkx object
  nx_graph <- as_networkx(res)
  print(nx_graph)

  # Convert to a list of data frames
  graph_dfs <- as.data.frame(nx_graph)
  print(graph_dfs$nodes)
  print(graph_dfs$edges)
}
}
