% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuzu.R
\name{kuzu_connection}
\alias{kuzu_connection}
\title{Create a Connection to a Kuzu Database}
\usage{
kuzu_connection(db)
}
\arguments{
\item{db}{A Kuzu database object, as returned by \code{kuzu_database()}.}
}
\value{
A Python object representing the connection to the Kuzu database.
}
\description{
Creates a new connection to an existing Kuzu database instance. Multiple
connections can be established to the same database.
}
\examples{
\dontrun{
db <- kuzu_database(":memory:")
conn <- kuzu_connection(db)
}
}
