% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kuzu_connection}
\alias{kuzu_connection}
\title{Create a connection to a Kuzu database}
\usage{
kuzu_connection(database)
}
\arguments{
\item{database}{Database object}
}
\value{
External pointer to the connection
}
\description{
Create a connection to a Kuzu database
}
