% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuz_load_data.R, R/kuzu_load_data.R
\name{kuzu_copy_from_csv}
\alias{kuzu_copy_from_csv}
\title{Load Data from a CSV File into a Kuzu Table}
\usage{
kuzu_copy_from_csv(conn, file_path, table_name, optionalcsvParameter = NULL)

kuzu_copy_from_csv(conn, file_path, table_name, optionalcsvParameter = NULL)
}
\arguments{
\item{conn}{A Kuzu connection object.}

\item{file_path}{A string specifying the path to the CSV file.}

\item{table_name}{A string specifying the name of the destination table in Kuzu.}

\item{optionalcsvParameter}{An optional parameter for CSV-specific configurations (e.g., delimiter, header).
Refer to Kuzu documentation for available options.}
}
\value{
This function is called for its side effect of loading data and does
not return a value.

This function is called for its side effect of loading data and does
not return a value.
}
\description{
Loads data from a CSV file into a specified table in the Kuzu database.

Loads data from a CSV file into a specified table in the Kuzu database.
}
\seealso{
\href{https://docs.kuzudb.com/import/csv/}{Kuzu CSV Import}

\href{https://docs.kuzudb.com/import/csv/}{Kuzu CSV Import}
}
