% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuz_load_data.R, R/kuzu_load_data.R
\name{kuzu_copy_from_json}
\alias{kuzu_copy_from_json}
\title{Load Data from a JSON File into a Kuzu Table}
\usage{
kuzu_copy_from_json(conn, file_path, table_name)

kuzu_copy_from_json(conn, file_path, table_name)
}
\arguments{
\item{conn}{A Kuzu connection object.}

\item{file_path}{A string specifying the path to the JSON file.}

\item{table_name}{A string specifying the name of the destination table in Kuzu.}
}
\value{
This function is called for its side effect of loading data and does
not return a value.

This function is called for its side effect of loading data and does
not return a value.
}
\description{
Loads data from a JSON file into a specified table in the Kuzu database.
This function also ensures the JSON extension is loaded and available.

Loads data from a JSON file into a specified table in the Kuzu database.
This function also ensures the JSON extension is loaded and available.
}
\seealso{
\href{https://docs.kuzudb.com/import/copy-from-json/}{Kuzu JSON Import}, \href{https://docs.kuzudb.com/extensions/json/}{Kuzu JSON Extension}

\href{https://docs.kuzudb.com/import/copy-from-json/}{Kuzu JSON Import}, \href{https://docs.kuzudb.com/extensions/json/}{Kuzu JSON Extension}
}
