% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuz_load_data.R, R/kuzu_load_data.R
\name{kuzu_copy_from_parquet}
\alias{kuzu_copy_from_parquet}
\title{Load Data from a Parquet File into a Kuzu Table}
\usage{
kuzu_copy_from_parquet(conn, file_path, table_name)

kuzu_copy_from_parquet(conn, file_path, table_name)
}
\arguments{
\item{conn}{A Kuzu connection object.}

\item{file_path}{A string specifying the path to the Parquet file.}

\item{table_name}{A string specifying the name of the destination table in Kuzu.}
}
\value{
This function is called for its side effect of loading data and does
not return a value.

This function is called for its side effect of loading data and does
not return a value.
}
\description{
Loads data from a Parquet file into a specified table in the Kuzu database.

Loads data from a Parquet file into a specified table in the Kuzu database.
}
\seealso{
\href{https://docs.kuzudb.com/import/parquet/}{Kuzu Parquet Import}

\href{https://docs.kuzudb.com/import/parquet/}{Kuzu Parquet Import}
}
