% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuzu_load_data.R
\name{kuzu_create_table_from_df}
\alias{kuzu_create_table_from_df}
\title{Create a Kuzu Table from a Data Frame}
\usage{
kuzu_create_table_from_df(conn, df, table_name, primary_key)
}
\arguments{
\item{conn}{A Kuzu connection object.}

\item{df}{A \code{data.frame} or \code{tibble} from which to infer the schema.}

\item{table_name}{A string specifying the name of the new table in Kuzu.}

\item{primary_key}{An optional string specifying the column to be used as the
primary key. If not provided, no primary key will be set.}
}
\value{
This function is called for its side effect of creating a table and
does not return a value.
}
\description{
Infers a schema from an R \code{data.frame} or \code{tibble} and creates a corresponding
NODE table in the Kuzu database.
}
\examples{
\donttest{
  conn <- kuzu_connection(":memory:")
  
  my_df <- data.frame(
    name = c("Alice", "Bob"),
    age = c(25L, 30L),
    height = c(1.75, 1.80),
    is_student = c(TRUE, FALSE),
    birth_date = as.Date(c("1999-01-01", "1994-05-15"))
  )
  
  kuzu_create_table_from_df(conn, my_df, "Person", primary_key = "name")
  
  # Now you can load data into the created table
  kuzu_copy_from_df(conn, my_df, "Person")
  
  result <- kuzu_execute(conn, "MATCH (p:Person) RETURN *")
  print(as.data.frame(result))
}
}
