% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kuzu_database}
\alias{kuzu_database}
\title{Create a Kuzu database}
\usage{
kuzu_database(db_path = NULL)
}
\arguments{
\item{db_path}{Path to database file, or ":memory:" for in-memory}
}
\value{
External pointer to the database
}
\description{
Create a Kuzu database
}
